* Mejorada notablemente la ventana de tramitación de citas. Ahora en el calendario se deshabilitan los días que el médico no pasa consulta y los días que podría pasar consulta pero tiene todas las horas ocupadas (y si se escribe a mano una fecha deshabilitada se muestra un mensaje, algo que antes no sucedía). De esta forma la ventana queda ex-ce-len-te. :) * Mejorado también el panel de consulta de beneficiarios, en el que ahora se muestra el nombre del médico asignado (no sólo su DNI) y el centro que tiene asignado, para que el beneficiario sepa a dónde debe dirigirse. * Mejorado el gestor de citas para tener en cuenta nuevas situaciones de error. * Se ha dividido la operación auxiliar 'consultarCitasMedico', que antes devolvía una matriz con dos tablas, en dos operaciones diferentes: 'consultarHorasCitas' y 'consultarCitasMedico'. * Creada una nueva operación auxiliar, 'consultarDiasCompletos', que se utiliza en el panel de tramitación de citas para poder desactivar los días que un médico pasa consulta pero tiene ya todas las citas ocupadas. * Añadido el evento BeneficiarioBuscado al panel JPBeneficiarioConsultar, para que se pueda saber cuándo se selecciona un beneficiario. * Mejorado el control de errores en el controlador del cliente para que no aparezcan mensajes de error poco descriptivos en la ventana principal. * Cambiada la clase y la tabla Sustitucion para que se indique la hora de inicio y final de las sustituciones. * Corregidos algunos errores en las pruebas de persistencia.

+++ practica1/app/servidorFrontend/src/dominio/conocimiento/IConstantes.java
@@ -8,6 +8,6 @@ public interface IConstantes {
 	
 	// La duración de las citas debe ser divisible entre 60 para
 	// que en una hora quepa un número entero de citas
-	public final static long DURACION_CITA = 15;
+	public final static int DURACION_CITA = 15;
 
 }
